Huge optimisation to the exogenous alignment quantification
exceRpt_Tools:
- Substantial improvement to the speed of ProcessExogenousAlignments.
Algorithm redesigned to traverse the taxonomy from bottom to top
(rather than top to bottom) to find the optimal alignment.  At each
iteration, the leaf node at which to start is selected as that which
has the most reads aligned to it.
- Introduced a new option in the bottom->top taxonomy walking that will
stop looking for new leaf nodes from which start then this leaf
represents less than some defined % of the total input reads.  A
sensible value for this is 1 in 100,000 reads (0.001%)